+++ ecs_logging/_utils.py
@@ -16,6 +16,7 @@
 # under the License.
 
 import json
+import functools
 
 try:
     import typing
@@ -152,6 +153,10 @@ def json_dumps(value):
     except KeyError:
         pass
 
+    json_dumps = functools.partial(
+        json.dumps, sort_keys=True, separators=(",", ":"), default=_json_dumps_fallback
+    )
+
     # Because we want to use 'sorted_keys=True' we manually build
     # the first three keys and then build the rest with json.dumps()
     if ordered_fields:
@@ -159,17 +164,34 @@ def json_dumps(value):
         # case the given values aren't strings (even though
         # they should be according to the spec)
         ordered_json = ",".join(
-            '"%s":%s' % (k, json.dumps(v, sort_keys=True, separators=(",", ":")))
+            '"%s":%s'
+            % (
+                k,
+                json_dumps(v),
+            )
             for k, v in ordered_fields
         )
         if value:
             return "{%s,%s" % (
                 ordered_json,
-                json.dumps(value, sort_keys=True, separators=(",", ":"))[1:],
+                json_dumps(value)[1:],
             )
         else:
             return "{%s}" % ordered_json
     # If there are no fields with ordering requirements we
     # pass everything into json.dumps()
     else:
-        return json.dumps(value, sort_keys=True, separators=(",", ":"))
+        return json_dumps(value)
+
+
+def _json_dumps_fallback(value):
+    # type: (Any) -> Any
+    """
+    Fallback handler for json.dumps to handle objects json doesn't know how to
+    serialize.
+    """
+    try:
+        # This is what structlog's json fallback does
+        return value.__structlog__()
+    except AttributeError:
+        return repr(value)