+++ tests/test_structlog_formatter.py
@@ -23,12 +23,18 @@ from .compat import StringIO
 import pytest
 
 
+class NotSerializable:
+    def __repr__(self):
+        return "<NotSerializable>"
+
+
 def make_event_dict():
     return {
         "event": "test message",
         "event.dataset": "agent.log",
         "log.logger": "logger-name",
         "foo": "bar",
+        "baz": NotSerializable(),
     }
 
 
@@ -47,7 +53,9 @@ def test_event_dict_formatted(time, spec_validator):
     formatter = ecs_logging.StructlogFormatter()
     assert spec_validator(formatter(None, "debug", make_event_dict())) == (
         '{"@timestamp":"2020-03-20T16:16:37.187Z","log.level":"debug",'
-        '"message":"test message","ecs":{"version":"1.6.0"},'
+        '"message":"test message",'
+        '"baz":"<NotSerializable>",'
+        '"ecs":{"version":"1.6.0"},'
         '"event":{"dataset":"agent.log"},'
         '"foo":"bar",'
         '"log":{"logger":"logger-name"}}'