+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed an issue in `StructlogFormatter` caused by a conflict with `event`
   (used for the log `message`) and `event.dataset` (a field provided by the
   `elasticapm` integration) ([#46](https://github.com/elastic/ecs-logging-python/pull/46))
+* Add default/fallback handling for json.dumps ([#47](https://github.com/elastic/ecs-logging-python/pull/47))
 
 ## 1.0.0 (2021-02-08)